32bit support

+++ binary-codec/codec.go
@@ -162,7 +162,7 @@ func EncodeForSigningBatch(json map[string]any) (string, error) {
 	// Create UInt32 for txIDs length
 	txIDsLengthType := &types.UInt32{}
 	txIDsLength := len(txIDsInterface)
-	if txIDsLength > math.MaxUint32 {
+	if uint32(txIDsLength) > uint32(math.MaxUint32) {
 		return "", ErrBatchTxIDsLengthTooLong
 	}
 	txIDsLengthBytes, err := txIDsLengthType.FromJSON(uint32(txIDsLength))
